// milo/ui/src/proto/go.chromium.org/luci/analysis/proto/v1/projects.pb.ts
export interface BugManagement {
  /**
   * The set of policies which control the (re-)opening, closure and
   * prioritization of bugs under the control of LUCI Analysis.
   */
  readonly policies: readonly BugManagementPolicy[];
  /** Monorail-specific bug filing configuration. */
  readonly monorail: MonorailProject | undefined;
}
// ...
export interface BugManagementPolicy {
  /**
   * A unique identifier for the bug management policy.
   *
   * Policies are stateful in that LUCI Analysis tracks which bugs have met the
   * activation condition on the policy (and not since met the deactivation
   * condition).
   *
   * Changing this value changes the identity of the policy and hence results in
   * the activation state for the policy being lost for all bugs.
   *
   * Valid syntax: ^[a-z]([a-z0-9-]{0,62}[a-z0-9])?$. (Syntax designed to comply
   * with google.aip.dev/122 for resource IDs.)
   */
  readonly id: string;
  /**
   * The owners of the policy, who can be contacted if there are issues/concerns
   * about the policy. Each item in the list should be an @google.com email
   * address. At least one owner (preferably a group) is required.
   */
  readonly owners: readonly string[];
  /**
   * A short one-line description for the problem the policy identifies, which
   * will appear on the UI and in bugs comments. This is a sentence fragment
   * and not a sentence, so please do NOT include a full stop and or starting
   * capital letter.
   *
   * For example, "test variant(s) are being exonerated in presubmit".
   */
  readonly humanReadableName: string;
  /**
   * The priority of the problem this policy defines.
   *
   * If:
   * - the priority of the bug associated with a rule
   *   differs from this priority, and
   * - the policy is activate on the rule (see `metrics`), and
   * - LUCI Analysis is controlling the priority of the bug
   *   (the "Update bug priority" switch on the rule is enabled),
   * the priority of the bug will be updated to match this priority.
   *
   * Where are there multiple policies active on the same rule,
   * the highest priority (of all active policies) will be used.
   *
   * For monorail projects, the buganizer priority will be converted to the
   * equivalent monorail priority (P0 is converted to Pri-0, P1 to Pri-1,
   * P2 to Pri-2, etc.) until monorail is turned down.
   */
  readonly priority: BuganizerPriority;
  /**
   * The set of metrics which will control activation of the bug-filing policy.
   * If a policy activates on a suggested cluster, a new bug will be filed.
   * If a policy activates on an existing rule cluster, the bug will be
   * updated.
   *
   * The policy will activate if the activation threshold is met on *ANY*
   * metric, and will de-activate only if the deactivation threshold is met
   * on *ALL* metrics.
   *
   * Activation on suggested clusters will be based on the metric values after
   * excluding failures for which a bug has already been filed. This is to
   * avoid duplicate bug filing.
   */
  readonly metrics: readonly BugManagementPolicy_Metric[];
  /**
   * Expanatory text of the problem the policy identified, shown on the
   * user interface when the user requests more information. Required.
   */
  readonly explanation: BugManagementPolicy_Explanation | undefined;
}
// ...
export interface MonorailProject {
  /**
   * The monorail project being described.
   * E.g. "chromium".
   */
  readonly project: string;
  /**
   * The prefix that should appear when displaying bugs from the
   * given bug tracking system. E.g. "crbug.com" or "fxbug.dev".
   * If no prefix is specified, only the bug number will appear.
   * Otherwise, the supplifed prefix will appear, followed by a
   * forward slash ("/"), followed by the bug number.
   * Valid prefixes match `^[a-z0-9\-.]{0,64}$`.
   */
  readonly displayPrefix: string;
}
// ...
export const BugManagement: MessageFns<BugManagement> = {
  encode(message: BugManagement, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    for (const v of message.policies) {
      BugManagementPolicy.encode(v!, writer.uint32(10).fork()).join();
    }
    if (message.monorail !== undefined) {
      MonorailProject.encode(message.monorail, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): BugManagement {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseBugManagement() as any;
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.policies.push(BugManagementPolicy.decode(reader, reader.uint32()));
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.monorail = MonorailProject.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): BugManagement {
    return {
      policies: globalThis.Array.isArray(object?.policies)
        ? object.policies.map((e: any) => BugManagementPolicy.fromJSON(e))
        : [],
      monorail: isSet(object.monorail) ? MonorailProject.fromJSON(object.monorail) : undefined,
    };
  },

  toJSON(message: BugManagement): unknown {
    const obj: any = {};
    if (message.policies?.length) {
      obj.policies = message.policies.map((e) => BugManagementPolicy.toJSON(e));
    }
    if (message.monorail !== undefined) {
      obj.monorail = MonorailProject.toJSON(message.monorail);
    }
    return obj;
  },

  create(base?: DeepPartial<BugManagement>): BugManagement {
    return BugManagement.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<BugManagement>): BugManagement {
    const message = createBaseBugManagement() as any;
    message.policies = object.policies?.map((e) => BugManagementPolicy.fromPartial(e)) || [];
    message.monorail = (object.monorail !== undefined && object.monorail !== null)
      ? MonorailProject.fromPartial(object.monorail)
      : undefined;
    return message;
  },
};

function createBaseBugManagementPolicy(): BugManagementPolicy {
  return { id: "", owners: [], humanReadableName: "", priority: 0, metrics: [], explanation: undefined };
}

export const BugManagementPolicy: MessageFns<BugManagementPolicy> = {
  encode(message: BugManagementPolicy, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.id !== "") {
      writer.uint32(10).string(message.id);
    }
    for (const v of message.owners) {
      writer.uint32(50).string(v!);
    }
    if (message.humanReadableName !== "") {
      writer.uint32(18).string(message.humanReadableName);
    }
    if (message.priority !== 0) {
      writer.uint32(24).int32(message.priority);
    }
    for (const v of message.metrics) {
      BugManagementPolicy_Metric.encode(v!, writer.uint32(34).fork()).join();
    }
    if (message.explanation !== undefined) {
      BugManagementPolicy_Explanation.encode(message.explanation, writer.uint32(42).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): BugManagementPolicy {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseBugManagementPolicy() as any;
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.id = reader.string();
          continue;
        }
        case 6: {
          if (tag !== 50) {
            break;
          }

          message.owners.push(reader.string());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.humanReadableName = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.priority = reader.int32() as any;
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.metrics.push(BugManagementPolicy_Metric.decode(reader, reader.uint32()));
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.explanation = BugManagementPolicy_Explanation.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): BugManagementPolicy {
    return {
      id: isSet(object.id) ? globalThis.String(object.id) : "",
      owners: globalThis.Array.isArray(object?.owners) ? object.owners.map((e: any) => globalThis.String(e)) : [],
      humanReadableName: isSet(object.humanReadableName) ? globalThis.String(object.humanReadableName) : "",
      priority: isSet(object.priority) ? buganizerPriorityFromJSON(object.priority) : 0,
      metrics: globalThis.Array.isArray(object?.metrics)
        ? object.metrics.map((e: any) => BugManagementPolicy_Metric.fromJSON(e))
        : [],
      explanation: isSet(object.explanation) ? BugManagementPolicy_Explanation.fromJSON(object.explanation) : undefined,
    };
  },

  toJSON(message: BugManagementPolicy): unknown {
    const obj: any = {};
    if (message.id !== "") {
      obj.id = message.id;
    }
    if (message.owners?.length) {
      obj.owners = message.owners;
    }
    if (message.humanReadableName !== "") {
      obj.humanReadableName = message.humanReadableName;
    }
    if (message.priority !== 0) {
      obj.priority = buganizerPriorityToJSON(message.priority);
    }
    if (message.metrics?.length) {
      obj.metrics = message.metrics.map((e) => BugManagementPolicy_Metric.toJSON(e));
    }
    if (message.explanation !== undefined) {
      obj.explanation = BugManagementPolicy_Explanation.toJSON(message.explanation);
    }
    return obj;
  },

  create(base?: DeepPartial<BugManagementPolicy>): BugManagementPolicy {
    return BugManagementPolicy.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<BugManagementPolicy>): BugManagementPolicy {
    const message = createBaseBugManagementPolicy() as any;
    message.id = object.id ?? "";
    message.owners = object.owners?.map((e) => e) || [];
    message.humanReadableName = object.humanReadableName ?? "";
    message.priority = object.priority ?? 0;
    message.metrics = object.metrics?.map((e) => BugManagementPolicy_Metric.fromPartial(e)) || [];
    message.explanation = (object.explanation !== undefined && object.explanation !== null)
      ? BugManagementPolicy_Explanation.fromPartial(object.explanation)
      : undefined;
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create(base?: DeepPartial<T>): T;
  fromPartial(object: DeepPartial<T>): T;
}